requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langgraph
chromadb
streamlit
pypdf
python-dotenv
openai
